+++ WebContent/WEB-INF/views/index.jsp
@@ -16,11 +16,11 @@
 		<!-- 로그인상태와 로그아웃상태일때 환영메세지, 관리자 일때 관리자페이지 링크가 보이게 한다. -->
 			<c:if test="${sessionScope.loginId != null}" >
 				<li>${sessionScope.loginId}님 환영합니다!!</li>
+				<li><a href="/buengbueng/logout.do">로그아웃</a></li>
 				<li><a href="/buengbueng/userInfoForm.do">회원 정보보기</a></li>
-				<li><a href="/buengbueng/userInfoFormUpdate.do">회원 정보수정</a></li>
 			</c:if>
 			<c:if test="${sessionScope.loginId == null}" >
-				<li><a href="">로그인</a></li>
+				<li><a href="/buengbueng/loginForm.do">로그인</a></li>
 				<li><a href="/buengbueng/userInfoSignForm.do">회원가입</a></li>
 				<li><a href="/buengbueng/userInfoSearchIdForm.do">아이디 찾기</a></li>
 				<li><a href="/buengbueng/userInfoSearchPwForm.do">비밀번호찾기</a></li>
