+++ WebContent/WEB-INF/views/userInfo/userInfoLogin.jsp
@@ -1,7 +1,7 @@
 <%@ page language="java" contentType="text/html; charset=UTF-8" %>
 
 <h1>로그인 테스트</h1>
-<form action="login.do">
+<form action="login.do" method="post">
 <div>id :<input type="text" name="id"></div>
 <div>pw :<input type="password" name="pw"></div>
 <input type="submit" value="로그인" >
