+++ src/sql/userInfoSQL.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!DOCTYPE sqlMap PUBLIC '-//ibatis.apache.org//DTD Sql Map 2.0//EN' 'http://ibatis.apache.org/dtd/sql-map-2.dtd'>
 <sqlMap namespace="test">
-	<typeAlias alias="userDto" type="login.user.bean.UserInfoDataBean" />
+	<typeAlias alias="userDto" type="login.user.bean.UserInfoDataDTO" />
 		
 	<!--  회원 삭제를 하기위한 패스워드 찾기 -->
 	<select id="checkPasswd" parameterClass="String" resultClass="String">
